Changes to be committed: 	renamed:    LSTM_baseline.ipynb -> baseline models/LSTM_baseline.ipynb 	renamed:    lstm-baseline-bonus.ipynb -> baseline models/lstm-baseline-bonus.ipynb 	renamed:    Untitled.ipynb -> final notebooks/Untitled.ipynb 	renamed:    aux-code-bonus.ipynb -> final notebooks/aux-code-bonus.ipynb 	renamed:    aux-code.ipynb -> final notebooks/aux-code.ipynb 	renamed:    hackathon-transformer-code.ipynb -> final notebooks/hackathon-transformer-code.ipynb 	renamed:    rockettransformer.ipynb -> final notebooks/rockettransformer.ipynb 	renamed:    time-series-forecasting-for-file.ipynb -> final notebooks/time-series-forecasting-for-file.ipynb 	renamed:    time-series-regression.ipynb -> final notebooks/time-series-regression.ipynb 	renamed:    aux_code.py -> getting started/aux_code.py 	renamed:    Confusion_matrix_LSTM.png -> images/Confusion_matrix_LSTM.png 	renamed:    Confusion_matrix_Transformer.png -> images/Confusion_matrix_Transformer.png 	renamed:    Sparse_categorical_accuracy.png -> images/Sparse_categorical_accuracy.png 	renamed:    input.png -> images/input.png 	renamed:    model_loss_LSTM.png -> images/model_loss_LSTM.png 	renamed:    model_loss_Transformer.png -> images/model_loss_Transformer.png 	renamed:    network.png -> images/network.png 	renamed:    network.svg -> images/network.svg 	renamed:    network_transformer_classification.png -> images/network_transformer_classification.png 	renamed:    regression.png -> images/regression.png 	renamed:    transformer.zip -> trained model/transformer.zip 	new file:   trained model/transformer_banana_muffin/keras_metadata.pb 	new file:   trained model/transformer_banana_muffin/saved_model.pb 	new file:   trained model/transformer_banana_muffin/variables/variables.data-00000-of-00001 	new file:   trained model/transformer_banana_muffin/variables/variables.index
