+++ images/Confusion_matrix_LSTM.png
