+++ images/model_loss_LSTM.png
